+++ Package.swift
@@ -1,17 +1,9 @@
 // swift-tools-version: 6.2
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
-import Foundation
 import PackageDescription
 
-let packageDirectory = URL(fileURLWithPath: #filePath).deletingLastPathComponent()
-let localCommanderPath = packageDirectory
-    .deletingLastPathComponent()
-    .appendingPathComponent("Commander")
-    .appendingPathComponent("Package.swift")
-let commanderDependency: Package.Dependency = FileManager.default.fileExists(atPath: localCommanderPath.path)
-    ? .package(path: "../Commander")
-    : .package(url: "https://github.com/steipete/Commander.git", from: "0.2.2")
+let commanderDependency: Package.Dependency = .package(url: "https://github.com/steipete/Commander", from: "0.2.2")
 
 let package = Package(
     name: "Tachikoma",